+++ src/drivers/amd_am79c973.h
@@ -86,7 +86,7 @@ namespace myos
 
             void Activate();
             int Reset();
-            common::uint32_t HandleInterrupt(common::uint32_t esp);
+            common::uint64_t HandleInterrupt(common::uint64_t esp);
 
             void Send(common::uint8_t* buffer, int count);
             void Receive();
